no image icons for new app, removed ununsed blur-chart image, rename browser wallet to defi wallet
