+++ src/assets/sass/theme.scss
@@ -668,7 +668,7 @@ code {
 
   .nfts-unlocks {
     font-size: 1.5rem;
-    margin-top: 1rem !important;
+    margin-top: 2rem !important;
     margin-bottom: 1rem !important;
   }
 }
