+++ src/components/TrailBlazerModal.tsx
@@ -4,7 +4,6 @@ import { FaCalendarCheck, FaChartBar, FaChessKnight, FaFlagCheckered, FaHandshak
 import { IoClose } from "react-icons/io5";
 import Modal from "react-modal";
 import { VerticalTimeline, VerticalTimelineElement } from "react-vertical-timeline-component";
-import imgBlurChart from "assets/img/blur-chart.png";
 import { Loader } from "components";
 import { useGetLoginInfo } from "@multiversx/sdk-dapp/hooks";
 import { IFrameModal } from "./iFrameModal";
@@ -248,7 +247,6 @@ export const TrailBlazerModal = ({
       <ModalBody>
         {!owned ? (
           <div className="d-flex flex-column align-items-center justify-content-center">
-            <img src={imgBlurChart} style={{ width: "90%", height: "auto" }} />
             <h4 className="mt-3 font-title">You do not own this Data NFT</h4>
             <h6>(Buy the Data NFT from the marketplace to unlock the data)</h6>
           </div>
