+++ src/libs/utils/constant.ts
@@ -1,6 +1,6 @@
 import iconBubbleMaps from "assets/img/expl-app-bubblemaps-icon.png";
 import iconTrailblazer from "assets/img/expl-app-trailblazer-icon.png";
-import iconInfrographics from "assets/img/expl-app-infographs-icon.png";
+import iconInfrographics from "assets/img/expl-app-infographics-icon.png";
 
 export const ERROR_CONNECT_WALLET = "Connect your wallet";
 export const ERROR_TRANSACTION_ONGOING = "A transaction is ongoing";
