+++ src/pages/Home.tsx
@@ -17,24 +17,18 @@ const HomePage = () => {
           <h3 className="c-app-market-title">App Marketplace</h3>
 
           <div className="row mt-3">
-            {APP_MAPPINGS.filter((app) =>
-              SUPPORTED_APPS.includes(app.routeKey)
-            ).map((item) => (
-              <div
-                key={item.routeKey}
-                className="col-12 col-md-6 col-lg-4 mb-3 d-flex justify-content-center c-app-tile"
-              >
+            {APP_MAPPINGS.filter((app) => SUPPORTED_APPS.includes(app.routeKey)).map((item) => (
+              <div key={item.routeKey} className="col-12 col-md-6 col-lg-4 mb-3 d-flex justify-content-center c-app-tile">
                 <div className="card shadow-sm border">
                   <div className="card-body p-3">
                     <div className="mb-4">
-                      <img className="data-app-image" src={item.img} />
+                      <Link to={returnRoute(item.routeKey)}>
+                        <img className="data-app-image" src={item.img} />
+                      </Link>
                     </div>
                     <h5 className="card-title">{item.appName}</h5>
                     <p className="card-text">{item.desc}</p>
-                    <Link
-                      to={returnRoute(item.routeKey)}
-                      className="btn btn-primary"
-                    >
+                    <Link to={returnRoute(item.routeKey)} className="btn btn-primary">
                       Launch App
                     </Link>
                   </div>
