+++ src/pages/ItheumTrailblazer.tsx
@@ -3,7 +3,7 @@ import { DataNft } from "@itheum/sdk-mx-data-nft";
 import { Address, SignableMessage } from "@multiversx/sdk-core/out";
 import { useGetLoginInfo } from "@multiversx/sdk-dapp/hooks";
 import { useGetLastSignedMessageSession } from "@multiversx/sdk-dapp/hooks/signMessage/useGetLastSignedMessageSession";
-import { useGetSignMessageInfoStatus } from '@multiversx/sdk-dapp/hooks/signMessage/useGetSignedMessageStatus';
+import { useGetSignMessageInfoStatus } from "@multiversx/sdk-dapp/hooks/signMessage/useGetSignedMessageStatus";
 import { useSignMessage } from "@multiversx/sdk-dapp/hooks/signMessage/useSignMessage";
 import { useNavigate, useParams } from "react-router-dom";
 import headerHero from "assets/img/custom-app-header-trailblazer.png";
@@ -25,8 +25,8 @@ export const ItheumTrailblazer = () => {
   const { targetNonce, targetMessageToBeSigned } = useParams();
   const { isPending: isSignMessagePending } = useGetSignMessageInfoStatus();
   const lastSignedMessageSession = useGetLastSignedMessageSession();
-  console.log('isSignMessagePending', isSignMessagePending);
-  console.log('lastSignedMessageSession', lastSignedMessageSession);
+  console.log("isSignMessagePending", isSignMessagePending);
+  console.log("lastSignedMessageSession", lastSignedMessageSession);
 
   const [itDataNfts, setItDataNfts] = useState<DataNft[]>([]);
   const [flags, setFlags] = useState<boolean[]>([]);
@@ -54,13 +54,13 @@ export const ItheumTrailblazer = () => {
       try {
         let signature = "";
 
-        if (lastSignedMessageSession && lastSignedMessageSession.status == 'signed' && lastSignedMessageSession.signature) {
+        if (lastSignedMessageSession && lastSignedMessageSession.status == "signed" && lastSignedMessageSession.signature) {
           signature = lastSignedMessageSession.signature;
         } else {
           let signSessions = JSON.parse(sessionStorage.getItem("persist:sdk-dapp-signedMessageInfo") ?? "{'signedSessions':{}}");
           signSessions = JSON.parse(signSessions.signedSessions);
           console.log("signSessions", signSessions);
-          
+
           // find the first 'signed' session
           for (const session of Object.values(signSessions) as any[]) {
             if (session.status && session.status == "signed" && session.signature) {
@@ -69,7 +69,7 @@ export const ItheumTrailblazer = () => {
             }
           }
         }
-        
+
         if (!signature) {
           throw Error("Signature is empty");
         }
@@ -198,7 +198,7 @@ export const ItheumTrailblazer = () => {
             <img className="img-fluid" src={headerHero} style={{ width: "100%", height: "auto" }} />
           </div>
           <div className="body">
-            <h4 className="my-3 text-center nfts-unlocks">Data NFTs that Unlock this App: {itDataNfts.length}</h4>
+            <h4 className="text-center nfts-unlocks">Data NFTs that Unlock this App: {itDataNfts.length}</h4>
 
             <div className="row mt-5">
               {itDataNfts.length > 0 ? (
